Travis now just runs 'rake'
Right now it runs `rake spec cucumber`. If we should ever update how
shoulda-matchers is tested in the future, there will be two places to
change this: Rakefile and .travis.yml. There is no reason we can't
simply use `rake` considering we have special logic in the Rakefile
which handles what happens when you call `rake` while BUNDLE_GEMFILE is
defined (which is the case on Travis, as it basically does what
appraisal is doing when you run tests locally).